+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and **Merged pull requests**. Critical items to know are:
 The versions coincide with releases on pip. Only major versions will be released as tags on Github.
 
 ## [0.0.x](https://github.com/flux-framework/flux-restful-api/tree/main) (0.0.x)
+ - Add better multi-user mode - running jobs on behalf of user (0.1.12)
  - Restore original rpc to get job info (has more information) (0.1.11)
  - Refactor of FLux Restful to use a database and OAauth2 (0.1.0)
  - Support for basic PAM authentication (0.0.11)